+++ audiodata.h
@@ -32,17 +32,17 @@ namespace KeyFinder {
 
     unsigned int getChannels() const;
     unsigned int getFrameRate() const;
-    float getSample(unsigned int index) const;
-    float getSampleByFrame(unsigned int frame, unsigned int channel) const;
-    float getSampleAtReadIterator() const;
+    double getSample(unsigned int index) const;
+    double getSampleByFrame(unsigned int frame, unsigned int channel) const;
+    double getSampleAtReadIterator() const;
     unsigned int getSampleCount() const;
     unsigned int getFrameCount() const;
 
     void setChannels(unsigned int newChannels);
     void setFrameRate(unsigned int newFrameRate);
-    void setSample(unsigned int index, float value);
-    void setSampleByFrame(unsigned int frame, unsigned int channels, float value);
-    void setSampleAtWriteIterator(float value);
+    void setSample(unsigned int index, double value);
+    void setSampleByFrame(unsigned int frame, unsigned int channels, double value);
+    void setSampleAtWriteIterator(double value);
     void addToSampleCount(unsigned int newSamples);
     void addToFrameCount(unsigned int newFrames);
 
@@ -60,11 +60,11 @@ namespace KeyFinder {
     AudioData* sliceSamplesFromBack(unsigned int sliceSampleCount);
 
   private:
-    std::deque<float> samples;
+    std::deque<double> samples;
     unsigned int channels;
     unsigned int frameRate;
-    std::deque<float>::const_iterator readIterator;
-    std::deque<float>::iterator writeIterator;
+    std::deque<double>::const_iterator readIterator;
+    std::deque<double>::iterator writeIterator;
   };
 
 }
