+++ fftadapter.h
@@ -34,11 +34,11 @@ namespace KeyFinder {
     FftAdapter(unsigned int frameSize);
     ~FftAdapter();
     unsigned int getFrameSize() const;
-    void setInput(unsigned int sample, float real);
+    void setInput(unsigned int sample, double real);
     void execute();
-    float getOutputReal(unsigned int bin) const;
-    float getOutputImaginary(unsigned int bin) const;
-    float getOutputMagnitude(unsigned int bin) const;
+    double getOutputReal(unsigned int bin) const;
+    double getOutputImaginary(unsigned int bin) const;
+    double getOutputMagnitude(unsigned int bin) const;
   protected:
     unsigned int frameSize;
     FftAdapterPrivate* priv;
@@ -49,9 +49,9 @@ namespace KeyFinder {
     InverseFftAdapter(unsigned int frameSize);
     ~InverseFftAdapter();
     unsigned int getFrameSize() const;
-    void setInput(unsigned int sample, float real, float imaginary);
+    void setInput(unsigned int sample, double real, double imaginary);
     void execute();
-    float getOutput(unsigned int bin) const;
+    double getOutput(unsigned int bin) const;
   protected:
     unsigned int frameSize;
     InverseFftAdapterPrivate* priv;
